crawlers/requirements.txt
pudb==2022.1.1
ipython==8.1.1
python-dateutil==2.8.2
pandas==1.4.3
mysqlclient==2.1.1
PyMySQL==1.0.2
pydantic==1.10.2
lxml==4.9.1
selenium==3.7
selenium-requests==1.3
beautifulsoup4==4.11.1
html5lib